"larutils

+++ buildScripts/buildUboone2.sh
@@ -14,14 +14,46 @@ echo "set qualifier: $SQUAL"
 echo "build type: $BUILDTYPE"
 echo "workspace: $WORKSPACE"
 
+# Parse qualifier parameter into base qualifier and label (if any).
+# Use hyphen as separator.
+
+basequal=''
+label=''
+quals=`echo $QUAL | tr '-' ' '`
+for q in $quals
+do
+  if [[ $q == e* || $q == c* ]]; then
+    basequal=$q
+  else
+    label=$q
+  fi
+done
+echo "basequal: $basequal"
+echo "label: $label"
+
+if [ x$basequal = x ]; then
+  echo "No base qualifier."
+  exit 1
+fi
+
 # Create area for biuld artifacts.
 rm -rf $WORKSPACE/copyBack
 mkdir -p $WORKSPACE/copyBack || exit 1
 
-# Check for supported combination of base qualifier and OS.
-if [[ `uname -s` == Darwin ]] && [[ $QUAL == e* ]]; then
-  echo "${QUAL} build not supported on `uname -s`"
-  echo "${QUAL} build not supported on `uname -s`" > $WORKSPACE/copyBack/skipping_build
+# Check for unsupported combinations of base qualifier, OS, and build label.
+if [[ `uname -s` == Darwin ]] && [[ $basequal == e* ]]; then
+  echo "${basequal} build not supported on `uname -s`"
+  echo "${basequal} build not supported on `uname -s`" > $WORKSPACE/copyBack/skipping_build
+  exit 0
+fi
+if [[ `uname -s` == Darwin ]] && [[ x$label == xpy2 ]]; then
+  echo "Python 2 build not supported on `uname -s`"
+  echo "Python 2 build not supported on `uname -s`" > $WORKSPACE/copyBack/skipping_build
+  exit 0
+fi
+if [[ `uname -s` == Linux ]] && [[ `lsb_release -rs` == 6* ]] && [[ x$label == x ]] && [[ $basequal != e17 ]] && [[ $basequal != c2 ]]; then
+  echo "Python 3 build not supported on SL6"
+  echo "Python 3 build not supported on SL6" > $WORKSPACE/copyBack/skipping_build
   exit 0
 fi
 
@@ -43,10 +75,17 @@ chmod +x buildFW
 echo
 echo "Begin build."
 echo
-./buildFW -t -b $QUAL -s $SQUAL $blddir $BUILDTYPE uboone-$VERSION || \
- { mv *.log $logdir
-   exit 1
- }
+if [ x$label = x ]; then
+  ./buildFW -t -b $basequal -s $SQUAL $blddir $BUILDTYPE uboone-$VERSION || \
+  { mv *.log $logdir
+    exit 1
+  }
+else
+  ./buildFW -t -b $basequal -s $SQUAL -l $label $blddir $BUILDTYPE uboone-$VERSION || \
+  { mv *.log $logdir
+    exit 1
+  }
+fi
 
 # Save log files.
